requirements.txt
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
contourpy==1.1.1
cryptography==42.0.5
cycler==0.12.1
DCNv3==1.0
DCNv4==1.0.0.post2
docutils==0.20.1
einops==0.7.0
filelock==3.9.0
fonttools==4.50.0
fsspec==2024.3.1
huggingface-hub==0.22.1
idna==3.6
importlib_metadata==7.1.0
importlib_resources==6.4.0
jaraco.classes==3.3.1
jaraco.context==4.3.0
jaraco.functools==4.0.0
jeepney==0.8.0
Jinja2==3.1.2
joblib==1.3.2
keyring==25.0.0
kiwisolver==1.4.5
markdown-it-py==3.0.0
MarkupSafe==2.1.3
matplot==0.1.9
matplotlib==3.7.5
mdurl==0.1.2
more-itertools==10.2.0
mpmath==1.3.0
networkx==3.0
nh3==0.2.15
ninja==1.11.1.1
numpy==1.24.1
nvidia-cublas-cu11==11.11.3.6
nvidia-cuda-cupti-cu11==11.8.87
nvidia-cuda-nvrtc-cu11==11.8.89
nvidia-cuda-runtime-cu11==11.8.89
nvidia-cudnn-cu11==8.7.0.84
nvidia-cufft-cu11==10.9.0.58
nvidia-curand-cu11==10.3.0.86
nvidia-cusolver-cu11==11.4.1.48
nvidia-cusparse-cu11==11.7.5.86
nvidia-nccl-cu11==2.19.3
nvidia-nvtx-cu11==11.8.86
onnx==1.16.0
opencv-python==4.9.0.80
packaging==24.0
pandas==2.0.3
pillow==10.2.0
pkginfo==1.10.0
protobuf==5.26.0
psutil==5.9.8
py-cpuinfo==9.0.0
pycparser==2.21
Pygments==2.17.2
pyloco==0.0.139
pyparsing==3.1.2
python-dateutil==2.9.0.post0
pytz==2024.1
pyutil==3.3.6
PyYAML==6.0.1
readme_renderer==43.0
requests==2.31.0
requests-toolbelt==1.0.0
rfc3986==2.0.0
rich==13.7.1
safetensors==0.4.2
scikit-learn==1.3.2
scipy==1.10.1
seaborn==0.13.2
SecretStorage==3.3.3
SimpleWebSocketServer==0.1.2
six==1.16.0
sympy==1.12
threadpoolctl==3.4.0
timm==0.9.16
torch==2.2.1+cu118
torchaudio==2.2.1+cu118
torchvision==0.17.1+cu118
tqdm==4.66.2
triton==2.2.0
twine==5.0.0
typing==3.7.4.3
typing_extensions==4.8.0
tzdata==2024.1
urllib3==2.2.1
ushlex==0.99.1
websocket-client==1.7.0
zipp==3.18.1
